4:35PM 29.1.2025 finished enter-room

+++ src/main/java/wordle/demo/rooms/Room.java
@@ -1,15 +1,15 @@
 package wordle.demo.rooms;
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
+import com.fasterxml.jackson.annotation.JsonManagedReference;
+import jakarta.persistence.*;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.PositiveOrZero;
 import lombok.Getter;
 import lombok.Setter;
+import wordle.demo.users.User;
 
 import java.time.LocalDateTime;
+import java.util.List;
 
 @Entity
 @Getter
@@ -18,7 +18,7 @@ public class Room {
 
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    private int id;
+    private Long id;
 
     @NotEmpty
     private String password;
@@ -27,4 +27,8 @@ public class Room {
     private int peopleAmount;
 
     private LocalDateTime emptyTime;
+
+//    @JsonManagedReference
+//    @OneToMany(mappedBy = "room", cascade = CascadeType.ALL, orphanRemoval = true)
+//    private List<User> users;
 }
